Struct waylyrics::config::Config

pub struct Config {

    pub player_sync_interval: String,
    pub lyric_update_interval: String,
    pub length_toleration: String,
    pub cache_lyrics: bool,
    pub theme: String,
    pub enable_filter_regex: bool,
    pub enable_local_lyric: bool,
    pub extract_translated_lyric: bool,
    pub filter_regexies: Vec<String>,
    pub lyric_search_source: Vec<String>,
    pub show_default_text_on_idle: bool,
    pub show_lyric_on_pause: bool,
    pub show_tray_icon: bool,
    pub player_name_blacklist: Vec<String>,
    pub player_identity_blacklist: Vec<String>,
    pub triggers: Triggers,
    pub qqmusic: QQMusicConfig,

}

Fields
player_sync_interval: String

the interval waylyrics updates position/metadata from player

lyric_update_interval: String

the interval waylyrics refreshes lyric labels

length_toleration: String

waylyrics matches lyrics with weights if (length-lyric_len).abs() < length_toleration, waylyrics set it’s weight as zero, mark it a best choice

cache_lyrics: bool

whether to cache lyrics note: persistenced lyric offset depends on this

theme: String

theme to load (.css)

enable_filter_regex: bool

if enabled, lyrics match one or more filter_regex will be hidden

enable_local_lyric: bool

if enabled, when waylyrics find xesam:url starts with file:// and no any other hint was matched, waylyrics will try to read the .lrc file has same filename

extract_translated_lyric: bool

if enabled, when waylyrics loads a local lyric file, it will try to extract translated lyric

filter_regexies: Vec<String>

hide lyric if it matches any of these regexies inspired by LyricX’s filter list

lyric_search_source: Vec<String>

avaliable options: 网易云音乐, QQ音乐, LRCLib

show_default_text_on_idle: bool

if enabled, waylyrics will set DEFAULT_TEXT on idle, otherwise it just show nothing

show_lyric_on_pause: bool

if enabled, waylyrics will show lyric on Pause

show_tray_icon: bool

whether to run tray-icon service

player_name_blacklist: Vec<String>

player with these name will be ignored

player_identity_blacklist: Vec<String>

player with these identity will be ignored

triggers: Triggers

shortcuts when focusing on waylyrics for global ones, please install the .desktop file also check trigger format at https://docs.gtk.org/gtk4/ctor.ShortcutTrigger.parse_string.html

qqmusic: QQMusicConfig

QQMusic config